Move command line classes to "CommandLine" namespace / folder
